Fix Xcode 6 / iOS 8 warnings for PushChat

+++ Samples/PushChat/PushChat/Containers/ChannelList.m
@@ -35,7 +35,7 @@
     
 }
 
--(int) channelCount
+-(NSUInteger) channelCount
 {
     return [self.channels count];
 }
